test(core): colocate split-mode regression tests per writer
Follow the repo's existing 1:1 source-to-test colocation pattern by
splitting the combined split-modes.test.ts into three files alongside
their writers (split-mode.test.ts, tags-mode.test.ts,
split-tags-mode.test.ts). Shared builder/output fixtures move to
test-utils/split-modes.ts so the per-writer files stay small.
No test logic changes - same six regression cases for #2309.
